Move setPageNumPaperSize util into extension commands
